+++ src/inspect_ai/model/__init__.py
@@ -95,6 +95,14 @@ from ._openai_convert import (
 from ._prompt import user_prompt
 from ._providers.providers import *
 from ._registry import modelapi
+from ._stream import (
+    StreamEvent,
+    StreamHandler,
+    StreamReasoningEvent,
+    StreamRetryEvent,
+    StreamTextEvent,
+    StreamToolCallEvent,
+)
 from ._trim import trim_messages
 
 __all__ = [
@@ -159,6 +167,12 @@ __all__ = [
     "StopCategory",
     "StopDetails",
     "StopReason",
+    "StreamEvent",
+    "StreamHandler",
+    "StreamReasoningEvent",
+    "StreamRetryEvent",
+    "StreamTextEvent",
+    "StreamToolCallEvent",
     "call_tools",
     "execute_tools",
     "ExecuteToolsResult",